Move morphology data outside the english folder
